build: fix installation and configuration
This commit fixes the installation and configuration of the libraries
and tools. wasmtime is now available.

The size of the image is huge but the idea is to create a base image
that only has the runtime and the s2i image itself will then just be
about taking the WebAssembly binary (wasm) or WebAssembly test format
(wat) and the configuration settings (fs, network settings) and setting
that up.

Still very much a work in progress.